[tensor] Power Tensor::apply() with c++17 parallel execution
This patch enables the speedup of Tensor::apply() with c++17 parallel
sequential and unsequential mode of execution.
This helps speed up all the activation funcations and various other
operations as well.
For reference, with this patch, unittest_tizen_capi takes 5.7sec to
finish compared to 7.2sec earlier. Similarily, unittest_nntrainer_models
takes 3.75sec compared to 4.88sec. The performance improvement depends
on the model architecture and operations used in it.

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>