[graph] Skip backwarding bug fix
Current implementation assumes that the graph is always linear and skips
the backwarding for contiguous set of layers at the start of the model
which should not do the backwarding.

This patch makes a fix to this issue but allowing layers in the middle
of the sorted graph to skip backwarding.

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>