[Fix] fix inconsistency when saving a model
When saving a model, if added order is different from sorted order,
loading a model with binary breaks some times

eg)
a -> b
  \
   v
   c

and layer was added a, b, c
it is being sorted to a, c, b

When this is saved, it is saved a, c, b.
When loading the graph from saved file it is loaded a, b, c.

This patch fixes this inconsistency happens when saving a model.

**Self evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test: [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Jihoon Lee <jhoon.it.lee@samsung.com>