[Trivial] Update gitignore file
**Update gitignore file**

The current build process involves downloading the necessary encoder,
ctre-unicode, and json files for running LLM from an external repository
which is not tracked within the nntrainer repo.

In order to prevent developers from accidentally uploading
these files upstream and to make the process more convenient,
we will be updating the gitignore file.

**Changes proposed in this PR:**
- Update gitignore file

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghak PARK <donghak.park@samsung.com>