[CI] Update labeler : Add READY2MERGE label
Currently, we have a git action that automatically labels as "Need Review" when the number of reviews is insufficient, and removes it when necessary. This automation allows developers to know which pull requests need review.

In addition, we want to automate the process of labeling pull requests that require merging, to make it more convenient for developers.

Resolves:
    #2463

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghak PARK <donghak.park@samsung.com>