[Application] fix some settings for LLaMA
The ENABLE_ENCODER2 option allows you to use specific tokenizer file on server.

However, this option does not fit the general user usage scenario because the tokenizer varies from every llama version to be distributed. In addition, the situation in which tokenizer is not used at all does not conform to the general usage scenario, too.

So, now the default setting is to always use the tokenizer of the path specified by the user.

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Seungbaek Hong <sb92.hong@samsung.com>