[Unittest] Add constant DerivativeLossLayer unittest
Currently, there is a layersemantics in the NNTrainer Unittest that verifies the behavior of the losslayer
but it does not include the case for the constant DerivativeLossLayer.

By simply filling in the parameters and passing them over, we can perform unit tests on basic operations
so I added this feature.

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghak PARK <donghak.park@samsung.com>