[Graph] Rename Variable & Function Names for clarity
currently, same term 'InPlace' means different things depending on context.

it could refer to a boolean type value indicating whether in-place operation is supported, or it could refer an enum type value representing the in-place type(type: None, Restriction, Non-Restriction).

To clarify this distinction, variables and functions related to the in-place type have been renamed from 'InPlace' to 'InPlaceType'.

Similarly, both functions for getting/setting whether it supports inplace (boolean) and those for getting/setting the type of inplace (enum) use the same name, "executeInPlace".

To clarify this, I changed the names as follows..

- set the InPlace flag (boolean) : 'setInPlace'.
- get the InPlace flag (boolean) : 'getInplace'.
- set the type of InPlace (enum) : 'setInPlaceType'
- get the type of InPlace (enum) : 'getInPlaceType.'

**Self evaluation:**
1. Build test:   [X]Passed [ ]Failed [ ]Skipped
2. Run test:     [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Seungbaek Hong <sb92.hong@samsung.com>