[Bugfix] Revert mixed precision case inplace set false
Currently, when using mixed precision mode with in-place set to true, the ninja test fails.

I think this failure is due to setting the in-place default value as true for all cases where it shouldn't have been.

Since loss layers and input layers haven't yet implemented non-FP32 operations, I've modified them so that they now specify in-place computations as false.

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghak PARK <donghak.park@samsung.com>