[FSU] Modify the condition of LoadTensors
Previously, more layers were loaded into memory than the value of
lookahead. I modified the code so that the correct number of layers
are loaded.

In addition, I removed the conditional statement so that the same
code is executed when the value of lookahead is 0 and when it is not.

Signed-off-by: SeoHyungjun <hyungjun.seo@samsung.com>